redirect all the admin dashboard into tasks directory
