--- smse_onboarding_main_project/admin_dashboard/__init__.py
@@ -1,1 +1,0 @@
-
