--- smse_onboarding_main_project/admin_dashboard/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-class AdminDashboardConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'admin_dashboard'
