--- smse_onboarding_main_project/admin_dashboard/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-
-app_name = 'admin_dashboard'
-
-urlpatterns = [
-    path('', views.admin_dashboard, name='admin_dashboard'),
-]
