+++ smse_onboarding_main_project/smse_onboarding/urls.py
@@ -19,6 +19,7 @@ from django.urls import include, path
 from django.views.generic.base import RedirectView
 from django.contrib.auth import views as auth_views
 from django.contrib.auth.decorators import login_required
+from tasks.views import login_redirect
 
 urlpatterns = [
     path('admin/', admin.site.urls),
@@ -26,8 +27,9 @@ urlpatterns = [
     path("home/", include("tasks.urls", namespace="tasks")),
     path('login/', auth_views.LoginView.as_view(
         template_name='login/login.html',
-        next_page='/admin-dashboard/'
+        next_page='/login-redirect/'
     ), name='login'),
+    path('login-redirect/', login_redirect, name='login_redirect'),
     path('logout/', auth_views.LogoutView.as_view(next_page='login'), name='logout'),
-    path('admin-dashboard/', include('admin_dashboard.urls', namespace='admin_dashboard')),
+    path('admin-dashboard/', RedirectView.as_view(url='/home/admin-dashboard/', permanent=True), name='admin_dashboard_redirect'),
 ]
