+++ smse_onboarding_main_project/tasks/static/admin_dashboard/README.txt
