+++ smse_onboarding_main_project/tasks/static/admin_dashboard/script.js
