+++ smse_onboarding_main_project/tasks/templates/admin_dashboard/admin_dashboard.html
